TD2 - Migration en Ligne/app_cdc/requirements.txt
mysql-connector-python==8.3.0
psycopg2-binary==2.9.9
